Don't show already imported movies as downloading in Movies index
